Script working
By putting the user response at the start instead that at the end the indices that I used to refer to the sentence tokens were off by one and the program responses had nothing to do with the user responses. I fixed that mistake and not the responses are at least pertinent to the user responses, even though the bot is far from "human".